signaling for notifiction api
This was intended to improve the performance of the most often used
request, the one to query for the existence of new notifications.
However, there is no performance gain to be measurable so that idea was
abandoned.

However, this at least fixes a bug the SQL based representers face when
no entries are returned by the selection of the resources listed in the
selection. Because it would result in an empty set, the expected json
would not be returned at all which then causes errors. The fix is
somewhat ugly but works.